Music Muted on Main Zoom Scene
Went through and locked the elevator music in an attempt to prevent automatic playing on the main scene.
Visible is set to "false" on the main zoom scenes, so the music should not be playing automatically upon switching to them.
There was also a caught typo in the What's New scene that prevented the music from being played at all.